add vm1 and vm2 pressure testing
