+++ cmd/tx-factory/src/lib.rs
@@ -1,3 +1,4 @@
+pub mod mocker;
 pub mod pressure_vm1;
 pub mod pressure_vm2;
 pub mod txn_generator;
